refactor(graph-gateway): split query context under client_query module
